# Conflicts:
#	calculator/src/main/java/com/market/banica/calculator/controller/CalculatorController.java
#	calculator/src/main/java/com/market/banica/calculator/dto/RecipeDTO.java
#	calculator/src/main/java/com/market/banica/calculator/model/Product.java
#	calculator/src/main/java/com/market/banica/calculator/service/CalculatorServiceImpl.java
#	calculator/src/main/java/com/market/banica/calculator/service/contract/CalculatorService.java
#	calculator/src/main/java/com/market/banica/calculator/service/grpc/AuroraClientSideService.java

+++ calculator/src/main/java/com/market/banica/calculator/service/contract/JMXServiceMBean.java
@@ -14,7 +14,7 @@ public interface JMXServiceMBean {
 
     void setProductQuantity(String recipeName, String ingredientName, int newQuantity);
 
-    int getProductQuantity(String recipeName, String ingredientName);
+    long getProductQuantity(String recipeName, String ingredientName);
 
     String getUnitOfMeasure(String productName);
